Refactor: memo를 이용한 리렌더링 방지
