+++ src/components/Table/TableNav/TableNavSelect/index.ts
@@ -1,0 +1,2 @@
+import TableNavSelect from './TableNavSelect';
+export { TableNavSelect };
